extract setupHooks
