+++ tests/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from "./sdk";
 export * from "./payloads";
+export * from "./setup";
