+++ tests/helpers/setup.ts
@@ -1,0 +1,24 @@
+import { beforeEachTruncateCollections } from "../hooks/collections";
+import { beforeAllCreateEngines } from "../hooks/engines";
+import { beforeEachLoadFixtures } from "../hooks/fixtures";
+import { useSdk } from "./sdk";
+
+export function setupHooks() {
+  const sdk = useSdk();
+
+  beforeAll(async () => {
+    await sdk.connect();
+    await beforeAllCreateEngines(sdk);
+  });
+
+  beforeEach(async () => {
+    await beforeEachTruncateCollections(sdk);
+    await beforeEachLoadFixtures(sdk);
+  });
+
+  afterAll(async () => {
+    sdk.disconnect();
+  });
+
+  return sdk;
+}
