+++ tests/modules/ingestion-pipeline/before-ingestion.test.ts
@@ -1,29 +1,11 @@
 import { ContainerAssetContent } from "../../../features/fixtures/application/assets/Container";
 
-import { beforeEachTruncateCollections } from "../../hooks/collections";
-import { beforeAllCreateEngines } from "../../hooks/engines";
-import { beforeEachLoadFixtures } from "../../hooks/fixtures";
-
-import { sendDummyTempPayloads, useSdk } from "../../helpers";
+import { sendDummyTempPayloads, setupHooks } from "../../helpers";
 
 jest.setTimeout(10000);
 
 describe("Ingestion Pipeline: before ingestion", () => {
-  const sdk = useSdk();
-
-  beforeAll(async () => {
-    await sdk.connect();
-    await beforeAllCreateEngines(sdk);
-  });
-
-  beforeEach(async () => {
-    await beforeEachTruncateCollections(sdk);
-    await beforeEachLoadFixtures(sdk);
-  });
-
-  afterAll(async () => {
-    sdk.disconnect();
-  });
+  const sdk = setupHooks();
 
   it("allows to add a new measure only to the asset", async () => {
     await sendDummyTempPayloads(sdk, [
